Use filter query for search within organization
The dataset search on the homepage uses a DisMax query, while the dataset search
on the organization does not. This produces wildly inconsistent results. In some
cases, no results appear when using search from the organization page. Using
a filter query for the organization produces better results. A true fix would be
making these queries the same (both DisMax or neither DisMax) and tuning search.